Pulled out ReorderLiterals transformer into separate file.
